fix: wc "no matching key" errors (#3316)
* fix: wc no matching key: keychain

This seems to be fixed with this PR:
https://github.com/WalletConnect/walletconnect-monorepo/pull/4139
so that is why I force a newer WC version since web3-onboard doesn’t have a release that uses it.

* fix: wc no matching key. session

This is an attempt to fix this WC error, but it doesn’t always work. Even more this error is reproducible only with 1inch.

Steps to reproduce the error:
1. Open your safe and login into an account
2. Open 1inch and select connect with walletconnect
3. On your safe, paste the link in the wc pairing field, but don’t approve
4. Refresh the page
5. Get a new pairing link from 1inch and try to pair again
6. You get the error

With the fix in this commit, we clean up on page unload this way the localstorage of walletconnect is left with an empty propsal and the next time the user tries to approve it should work. This is only a workaround. I think that this is a bug in walletconnect and they should better take care of their localstorage.