fix: cannot connect with WC to Opensea (#3437)
* fix: cannot connect with WC to Opensea

If the user is trying to connect with WC to Opensea and the user is on a chain other than mainnet the first connection attempt fails. The second one succeeds. This is because when opensea sees the list of optional chains, after establishing a session it sends a request, but the request is for the optional chain and not for the required one and we end up blocking it. By giving up on the optional namespace Opensea sends the request to the requried chain.

We then immediately update the session with the optional namespaces - this way switching accounts works fine for dapps and we don’t disconnect when we switch networks.