Feat: Native Swaps (#3231)
This PR ads a convenient swap feature directly in the Safe Interface. The swap feature can be accessed from the dashboard, assets page or directly from the navbar. Under the hood this is integration with cowswap. We decoded the cowswap transactions to make a swap more convenient and secure & push notifications about the status of a swap appear in the safe interface. 

---------

Co-authored-by: Usame Algan <usame@safe.global>
Co-authored-by: Usame Algan <5880855+usame-algan@users.noreply.github.com>
Co-authored-by: Aaron Cook <aaron@safe.global>
Co-authored-by: Tanya <93707525+TanyaEfremova@users.noreply.github.com>
Co-authored-by: James Mealy <james@safe.global>
Co-authored-by: Manuel Gellfart <manu@safe.global>
Co-authored-by: katspaugh <381895+katspaugh@users.noreply.github.com>