fix: swap token selection flickers (#4006)
Tokens can end up having the same token abbreviation, but they always
have different token addresses. If the user select a token that has
the same abbreviation as another token in the list the widget starts
flickering between the selected tokens and resets to WETH for input
token and no selected output token. By opting to use addresses instead
the problem seems to be resolved.