Refactor: generate terms page from md file (#4183)
* fix: display last updated version in cookie banner

The cookie banner now shows the last updated version. In addition to this
it is no longer necessary to update the version number inside of the
store slice. We can control it from the frontmatter of the terms.md doc.

Whenever the version gets updated the cookie banner is going to pop up.