feat(mobile): move the accounts bottom sheet to separate screen
The current implementation of the dropdown component had one drawback:
it could only be closed from the screen on which the dropdown was
rendered.

Imagine the following situation.
“home screen” -> user opens “sheet modal” -> open “scan safe” screen
-> open “import overview” screen -> open “signers screen” ->
open “import screen” -> import

The last step in the import flow is to save the “safe wallet” in the
safes store slice. Once this is done the expected behavior would be to
send the user back to the home screen with the newly saved safe as
active. To do this on the last step in the import flow we can do:
router.dismissTo(“/“) -> this will navigate the user to the “scan safe”
screen (the first screen in the stack).
Now to close it and return to the previous stack we do
router.back() unfortunately the user is presented with the “sheet modal”
instead of the “home” screen. Unfortunately form the
“import” screen we cannot control the state of the “sheet modal”.

To go around this problem, I’ve refactored the code. Now the
“sheet modal” is a separate “accounts sheet” screens. This way it is
part of the navigation and we can dismiss it through the router.

In the above case, since we do:
router.dismissTO(“/“) -> this brings us to the “scan safe” screen
router.back() -> brings us to the “accounts sheet” screen
router.back() -> brings us to the “home” screen