fix(mobile): multiple UI inconsistencies (#5351)
* fix(mobile): identicon is rounded everywhere

The identicon in our design is rounded everywhere. We have a rounded prop, which is always set to false and there were a lot of places where we forgot to set rounded=true. Because of this I’m switching the syntax.

* fix: UI inconsistencies

When a screen is rendered inside of a modal the continue button was not visible due to bugs in React-Natives new Architecture and react-native-screens.
I tried multiple workarounds and right now it seems to display ok in most cases.