refactor(utils): move signTxServiceMessage function back to web package (#5533)
* refactor(utils): move signTxServiceMessage back to web

This functions needs signTypeData from the web utils/web3 helper. Unfortunately if I move the helper to the utils package it drags with itself the typedData dependency from the safe-gateway-typescript-sdk.
If I don’t move the function the mobile app can’t be build because the @safe-global/web package has a stream dependency from nodejs which cannot be resolved by the metro bundler. Until we ditch the safe-gateway-typescript-sdk this type will live in apps/web package.

* fix(utils): tests were failing

* chore(ci): add a unit test workflow for utils package

* fix(utils): ci fails