fix(mobile): build failure (#6066)
* fix(mobile): production build warning

should fix the following warning:
warning:  The file reference for "NotificationService.swift" is a member of multiple groups ("NotifeeNotificationServiceExtension" and "NotifeeNotificationServiceExtension"); this indicates a malformed project.  Only the membership in one of the groups will be preserved (but membership in targets will be unaffected).  If you want a reference to the same file in more than one group, please add another reference to the same path.

because the NotificationService.swift is both in the files_to_add  and in the extension_service_file due to the way the expo script is written the file ends up twice in the xcodeproj file

* fix(mobile): production build fails