fix(web): EIP-7702 accounts are treated as smart accounts (#6235)
Our isSmartContract check returns true whenever getCode returns anything other than an empty data. For EIP-7702 accounts we do have data which is prefixed with 0xef0100. Those accounts are both smart accounts and EOAs. The user is still able to use them as EOAs.
With this commit we check if the data contains 0xef0100 and treat such accounts as EOA when it comes to any wallet functions.