fix(mobile): notification badge was not cleared (#6576)
The notification badge was only cleared when the user would click on the push notification. If the user would open the app, by pressing the icon the badge won’t be cleared. Now when the user navigatese to the pending screen or history we will clear the badge. Until we have the notification center I think that this is an acceptable solution.