fix(mobile): flicker when scrolling in history (#6811)
I think that I re-introduced the bug when I fixed the missing skeleton loader in the footer. It seems that the flicker was due to the listHeaderComponent briefly rendering when we were scrolling down which would push the content down and then it will jump again up. I think that right now we don’t need ListHeaderComponent at all. I was trying to have 2 way infinite scroll, but it is not supported by RTK right now. Removing the ListHeaderComponent all together seems to have solved the flicker.