chore: unify prettier setup across repo (#6849)
* chore: unify prettier setup across repo

there were slight inconsistencies in the prettier setup across the repo which lead to annoying problems. Running prettier from the root would modify files in subdirectories. But if you would run prettier from those subdirecotry -> the same files that were previously modified wouldn’t be treated as a problem.

Now every prettier command across the repo is loading the same root prettier config. In addition to this lintstaged should now only run on the staged files and not on the complete code base.

* fix: prettier

it turns out that the node_modules/.bin/prettier that we end up with is not from our set prettier v3.6.2, but rather from jest-snapshot-prettier. It sets:
  bin:
    prettier: bin-prettier.js

So, when we run `yarn prettier` - yarn correctly resolves with v3 and our config. But when we run it through lint-staged since we just call prettier - we end up with the jest-snapshot-prettier prettier binary… This would then reformat the files again and will change the changes made by prettier v3…

* fix: .prettierignore